fixed sharply looking damage blob and camera offset when time is big (toolset v120 problem)

+++ codemp/cgame/cg_local.h
@@ -957,7 +957,7 @@ typedef struct {
 	int			weaponAnimationTime;
 
 	// blend blobs
-	float		damageTime;
+	int			damageTime;
 	float		damageX, damageY, damageValue;
 
 	// status bar head
